Use `runBlocking` close to the source
Initially, I wanted to move `runBlocking` wrappers away from source, to the presentation layer, as this would be easier to fix and manage. Unfortunately, it's not possible as tests rely on synchronous nature of these calls. So for now, I'm moving the `runBlocking` wrappers as close to the source as possible