Fix: Replace payload with result on delete variations for product
It looks like as part of 1df49aa69f4f15d1bc702f33746cc7003a62e790 an
oversight occurred where, instead of 'result.remoteProductId',
'payload.remoteProductId' got used.

I am not sure if that was deliberate or not, and since there was no
specific comment about that, it is better to treat that as an oversight.
This commit serves as an explicit nudge to reviewers, and the original
auhtor, to verify this change with that extra 'fix' commit.