Refactor: Use local/remote instead of int/long ids for product variation
This follows the same pattern applied to 'WCProductModel' and it seems
that this pattern is the one mostly used. This pattern is similar to
using a 'value' class to create a more domain-specific type instead of
primitive types to help with data class construction clarity.

Also, related to: 316ef5fae66bb639beaf1e47342e6c26c0875953