Db: Remove unused 'PostModelXyz' tables/code from 'wp-fluxc'
List of tables:
- PostModel
- PostUploadModel

Also, as part of this change, the below removal occurred:
- SiteStore: SiteStore.handleFetchedSitesWPComRest(...) ->
    SiteSqlUtils.removeWPComRestSitesAbsentFromList(...) ->
      PostSqlUtils.getSiteHasLocalChanges(...) logic got removed.
- UploadStore: UploadStore.handleMediaUploaded/Cancel/UpdateMedia ->
    UploadStore.cancelPost(...) logic got removed.
- PostListDescriptor: Both ListItemSqlUtilsTest/ListSqlUtilsTest test
suites got completely removed.

FYI: All above removal happened so that the project could compile, and
also, so that no 'post' package references remain on this project.

PS: The only exception to this above is 'PostWPComRestResponse' and its
'post' package. This is due to the fact that this response data class is
being used by the 'ProductRestClient' class to both,
'fetchProductPassword(...)' and 'updateProductPassword(...)'.