Fix: Add back 'FetchMediaListPayload' ctor to fix upload method crash
This fixes the 'java.lang.NoSuchMethodError' on
'MediaStore$FetchMediaListPayload' for when a user tries to add a new
product photo using the 'WordPress media library' option.

FYI: This crash occurs because the 'WordPress-MediaPicker-Android'
library is using an older version of the 'WordPress-FluxC-Android'
library, which is using this specific 'FetchMediaListPayload'
constructor.