Use a single model for both PurchasedLabelDTO and ShippingLabelDTO
Previously, there were two separate models due to the "id" field being either a Long or a String. I’ve now combined them into one model by defining the id as a String. This allows us to parse the value as a string regardless of whether it’s originally a number or a string.