Remove deprecated and unnecessary use of announceForAccessibility() call
announceForAccessibility is deprecated now and is generally used for dynamic announcements (like a status update) rather than for labeling static content like a header. While it works, it's not the standard way to mark a view as a heading.

A view marked as a heading should have textual content that the screen reader can announce as the heading's title. So adding announceForAccessibility() is redundant.