Migration: Integrate automattic-encryptedlogging 1.0.0
This change includes but it is not limited to the below:
1. Everything 'lazysodium', 'jna' and 'encryptedlogging' being removed
from the 'fluxc' module, including related tests.
2. On the 'WooCommerce' module, this newest version of
'encryptedlogging', along with its latest API changes, updates the
previous implementation. Related tests are updated as well.
3. The 'ObserveEncryptedLogsUploadResult' is now completely removed as
no longer relevant (ish), no APIs to call.
4. This 'encryptedLogging.resetUploadStates()' logic is moved from the
now deleted 'ObserveEncryptedLogsUploadResult' class and into
'UploadEncryptedLogs' class. This is done because:
  - Both the below calls were anyway being called sequentially with
    'AppInitializer':
    - observeEncryptedLogsUploadResults()
    - uploadEncryptedLogs()
  - Other projects, like DOAndroid and PCAndroid, are doing exactly the
    same.