Analysis: Suppress deprecated warning
Warning Message: "'static fun startActivity(context: @NonNull() Context,
intent: @NonNull() Intent, options: @Nullable() Bundle?): Unit' is
deprecated. Deprecated in Java"

FYI: This warning start appearing because adding this new version of
'encrypted-logging' into this project brings with it a lot of transitive
dependencies that are on a higher version that that the existing
explicitly set dependency on this project.