Removes or replace deprecated usages of announceForAccessibility
We were using announceForAccessibility in CardReaderUpdateDialogFragment to indicate talkback service to announce 2 things:
1. Announce card reader initiating
2. Automatically report upgrade progress everytime the progress percentage is updated.

For the first case, we don't need to explicitly announce the updated is starting as this is something talkback service will do automatically.

For the second case, non deprecated announceAccessibilityChange was added which defines a live region to report progress to talk back service.