Split the delete password functions
The delete functions are used for two scenarios: logout, or a password conflict. The latter requires fetching the UUID, and will cause issues if it's used for Jetpack access, as it will try fetching the UUID even if we didn't create an app password.
Splitting the roles makes things easier to follow.