Fix: Revert enqueue sending encrypted logs back to blocking
FYI: This is done because we've identified a regression where the actual
encrypted log, although seen as sent, is not actually ending-up being
uploaded into our systems. Thus, when someone then tries to browse it,
based on it UUID, they will get: "Log File Not Found.Invalid Log File"

For more info see: p1759832867149029-slack-CGPNUU63E

PS: The PR that introduced this regression is #14407, which brought-in
some app logs improvements.