Refactor: Make revenue range id not empty by default
It seems that 'revenueRangeId' it totally app driven, and not REST API
driven, and as such, can be completely manipulated, effectively becoming
a non-empty value at all times.

With this change, we are making sure that when 'revenueRangeId' comes
from places like 'Widget' or 'Wear', we are adding the necessary missing
information to uniquely identify this revenue range id. We are using
this '<Medium><StartDate><EndDate>' pattern.

FYI: This extra 'Medium' data was added to easily identify the medium.
With or without 'Medium' this would have worked fine anyway.

PS: Where needed, 'Medium' is either 'Widget' or 'Wear'.

------------------------------------------------------------------------

PR Comment: https://github.com/woocommerce/woocommerce-android/pull/
14680#discussion_r2398949681