Re-add `.configure-files` to `.gitignore`
To cover the legacy case of someone switching to an old branch that didn't have `git-crypt` yet, running `configure_apply` on that old branch, then switching back to a branch that used `git-crypt`. In that case there's a risk that some files in `.configure-files` that would contain decrypted secrets would still be around (if the used didn't run `git clean`) and be accidentally commited to the remote.