Remove setting `setLastVersionWithAnnouncement` after the `getLatestFeatureAnnouncement(fromCache = false)`
This was an optimization that would save us from an additional HTTP request on the app start.

It's not necessary though, and brings a risk of not ever showing announcement if `getLatestFeatureAnnouncement` fails. Addressing "failed request" case would need a refactor to distinguish between `null` (no announcements) and `failure` (http request failed).

So we keep things as they're now on `trunk` - each time on the app start the HTTP request is triggered.